requirements.txt
pandas
simple_salesforce
python-dotenv
